feat: add post convertion step to handel inheritance
The post convertion step looks at all the definitions created
by the single file convertion step. It looks up all the parent
definitions given by extends. When a Definition with a parent
is found the parent Definition is looked up in the set of all
Definitions and the fields of the parent definition is copied
into the start of the child definition.

This handles one level of inheritance.