refactor: remove util package from canonical name factory
The util package is not auto imported (like the lang package) in
Java. Thus whenever a model contains a List or a Map the model will also
contain a import statement for the type. Therefor we do not have to
include a special mapping for Map and List types in the
CanonicalNameFactory.